+++ Fast-Order.sln
@@ -69,6 +69,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Postech8SOAT.FastOrder.Infr
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Postech8SOAT.FastOrder.Integration.Tests", "test\Postech8SOAT.FastOrder.Integration.Tests\Postech8SOAT.FastOrder.Integration.Tests.csproj", "{281D63FD-CE85-4D8D-84D9-1A2D12EB9F08}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Postech8SOAT.FastOrder.Domain.BDD.Tests", "test\Postech8SOAT.FastOrder.Domain.BDD.Tests\Postech8SOAT.FastOrder.Domain.BDD.Tests.csproj", "{32DE2707-D714-49A9-A265-61C672AAECB5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -131,6 +133,10 @@ Global
 		{281D63FD-CE85-4D8D-84D9-1A2D12EB9F08}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{281D63FD-CE85-4D8D-84D9-1A2D12EB9F08}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{281D63FD-CE85-4D8D-84D9-1A2D12EB9F08}.Release|Any CPU.Build.0 = Release|Any CPU
+		{32DE2707-D714-49A9-A265-61C672AAECB5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{32DE2707-D714-49A9-A265-61C672AAECB5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{32DE2707-D714-49A9-A265-61C672AAECB5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{32DE2707-D714-49A9-A265-61C672AAECB5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -157,6 +163,7 @@ Global
 		{85348F60-50A9-4961-93F1-B9E34B4605C5} = {8CD53F91-3587-4608-8006-AFDF492C956D}
 		{DBF601BA-6527-4074-A761-ACA7FFA5CBA0} = {8CD53F91-3587-4608-8006-AFDF492C956D}
 		{281D63FD-CE85-4D8D-84D9-1A2D12EB9F08} = {4A3D97B3-2F7C-41B7-87CD-E376FD88B288}
+		{32DE2707-D714-49A9-A265-61C672AAECB5} = {4A3D97B3-2F7C-41B7-87CD-E376FD88B288}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {290B78A1-7B91-4DC6-AB6E-64D81B6C5F7D}
