+++ test/Postech8SOAT.FastOrder.Domain.BDD.Tests/Feature/Categoria.feature
