+++ test/Postech8SOAT.FastOrder.Domain.BDD.Tests/Feature/Categoria.feature.cs
@@ -10,7 +10,7 @@
 // ------------------------------------------------------------------------------
 #region Designer generated code
 #pragma warning disable
-namespace Postech8SOAT.FastOrder.Domain.Tests.Feature
+namespace Postech8SOAT.FastOrder.Domain.BDD.Tests.Feature
 {
     using TechTalk.SpecFlow;
     using System;
@@ -19,7 +19,7 @@ namespace Postech8SOAT.FastOrder.Domain.Tests.Feature
     
     [System.CodeDom.Compiler.GeneratedCodeAttribute("TechTalk.SpecFlow", "3.9.0.0")]
     [System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
-    public partial class GerenciamentoDeClientesFeature : object, Xunit.IClassFixture<GerenciamentoDeClientesFeature.FixtureData>, System.IDisposable
+    public partial class GerenciamentoDeCategoriasFeature : object, Xunit.IClassFixture<GerenciamentoDeCategoriasFeature.FixtureData>, System.IDisposable
     {
         
         private static TechTalk.SpecFlow.ITestRunner testRunner;
@@ -28,10 +28,10 @@ namespace Postech8SOAT.FastOrder.Domain.Tests.Feature
         
         private Xunit.Abstractions.ITestOutputHelper _testOutputHelper;
         
-#line 1 "Cliente.feature"
+#line 1 "Categoria.feature"
 #line hidden
         
-        public GerenciamentoDeClientesFeature(GerenciamentoDeClientesFeature.FixtureData fixtureData, Postech8SOAT_FastOrder_Domain_Tests_XUnitAssemblyFixture assemblyFixture, Xunit.Abstractions.ITestOutputHelper testOutputHelper)
+        public GerenciamentoDeCategoriasFeature(GerenciamentoDeCategoriasFeature.FixtureData fixtureData, Postech8SOAT_FastOrder_Domain_BDD_Tests_XUnitAssemblyFixture assemblyFixture, Xunit.Abstractions.ITestOutputHelper testOutputHelper)
         {
             this._testOutputHelper = testOutputHelper;
             this.TestInitialize();
@@ -40,8 +40,8 @@ namespace Postech8SOAT.FastOrder.Domain.Tests.Feature
         public static void FeatureSetup()
         {
             testRunner = TechTalk.SpecFlow.TestRunnerManager.GetTestRunner();
-            TechTalk.SpecFlow.FeatureInfo featureInfo = new TechTalk.SpecFlow.FeatureInfo(new System.Globalization.CultureInfo("pt-BR"), "Feature", "Gerenciamento de Clientes", "  Como um sistema de gerenciamento de clientes\r\n  Eu quero garantir que os client" +
-                    "es sejam criados, alterados e validados corretamente", ProgrammingLanguage.CSharp, featureTags);
+            TechTalk.SpecFlow.FeatureInfo featureInfo = new TechTalk.SpecFlow.FeatureInfo(new System.Globalization.CultureInfo("pt-BR"), "Feature", "Gerenciamento de Categorias", "  Para garantir a consistência dos dados\r\n  Como um desenvolvedor\r\n  Quero valida" +
+                    "r as regras de negócio ao criar uma categoria", ProgrammingLanguage.CSharp, featureTags);
             testRunner.OnFeatureStart(featureInfo);
         }
         
@@ -81,16 +81,16 @@ namespace Postech8SOAT.FastOrder.Domain.Tests.Feature
             this.TestTearDown();
         }
         
-        [Xunit.SkippableFactAttribute(DisplayName="Criar um cliente válido")]
-        [Xunit.TraitAttribute("FeatureTitle", "Gerenciamento de Clientes")]
-        [Xunit.TraitAttribute("Description", "Criar um cliente válido")]
-        public void CriarUmClienteValido()
+        [Xunit.SkippableFactAttribute(DisplayName="Deve criar categoria com sucesso")]
+        [Xunit.TraitAttribute("FeatureTitle", "Gerenciamento de Categorias")]
+        [Xunit.TraitAttribute("Description", "Deve criar categoria com sucesso")]
+        public void DeveCriarCategoriaComSucesso()
         {
             string[] tagsOfScenario = ((string[])(null));
             System.Collections.Specialized.OrderedDictionary argumentsOfScenario = new System.Collections.Specialized.OrderedDictionary();
-            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Criar um cliente válido", null, tagsOfScenario, argumentsOfScenario, featureTags);
-#line 6
-  this.ScenarioInitialize(scenarioInfo);
+            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Deve criar categoria com sucesso", null, tagsOfScenario, argumentsOfScenario, featureTags);
+#line 19
+this.ScenarioInitialize(scenarioInfo);
 #line hidden
             if ((TagHelper.ContainsIgnoreTag(tagsOfScenario) || TagHelper.ContainsIgnoreTag(featureTags)))
             {
@@ -99,20 +99,17 @@ namespace Postech8SOAT.FastOrder.Domain.Tests.Feature
             else
             {
                 this.ScenarioStart();
-#line 7
-    testRunner.Given("que o CPF do cliente é \"765.273.200-00\"", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Dado ");
+#line 20
+    testRunner.Given("que o nome da categoria é \"Teste\"", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Dado ");
 #line hidden
-#line 8
-    testRunner.And("o nome do cliente é \"João da Silva\"", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "E ");
+#line 21
+    testRunner.And("a descrição da categoria é \"Teste\"", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "E ");
 #line hidden
-#line 9
-    testRunner.And("o email do cliente é \"joao@example.com\"", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "E ");
+#line 22
+    testRunner.When("eu criar uma categoria", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Quando ");
 #line hidden
-#line 10
-    testRunner.When("eu criar o cliente", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Quando ");
-#line hidden
-#line 11
-    testRunner.Then("o cliente deve ser válido", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Então ");
+#line 23
+    testRunner.Then("a categoria deve ser criada com sucesso", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Então ");
 #line hidden
             }
             this.ScenarioCleanup();
@@ -125,12 +122,12 @@ namespace Postech8SOAT.FastOrder.Domain.Tests.Feature
             
             public FixtureData()
             {
-                GerenciamentoDeClientesFeature.FeatureSetup();
+                GerenciamentoDeCategoriasFeature.FeatureSetup();
             }
             
             void System.IDisposable.Dispose()
             {
-                GerenciamentoDeClientesFeature.FeatureTearDown();
+                GerenciamentoDeCategoriasFeature.FeatureTearDown();
             }
         }
     }
